chore(analysis): upgrade PHPStan to 2.x, baseline the delta, fix surfaced bugs
Bump phpstan/phpstan to ^2.0 with phpstan-phpunit, phpstan-strict-rules and
phpstan-deprecation-rules on their 2.x majors. The level-6 gate stays green:
the 2.x engine's new findings are frozen in phpstan-baseline.neon (111
identifier-tagged entries — burn down, never grow). Drop the removed
checkGenericClassInNonGenericObjectType option and add the missing
phpstan:di script (src/Container, level 8) that types:validate referenced.

Real bugs the sharper engine surfaced, fixed with coverage:
- QueryBuilder/WhereClause gain orWhereRaw(); the ORM relation-count path
  dispatched to it for the 'or' boolean, so orHas()/orDoesntHave()/
  orWhereHas()/orWhereDoesntHave() fataled (pinned in OrWhereRawTest)
- RedisCacheDriver::zadd() no longer ArgumentCountErrors on an empty map
- integer array indices in median/formatBytes math, a malformed
  @SuppressWarnings docblock, a dead is_bool branch on an int return

Level-8 debt catalog re-surveyed under 2.x: 839 errors (was 914 under 1.12).

+++ src/DTOs/PaginatedResponseDTO.php
@@ -127,7 +127,7 @@ class PaginatedResponseDTO
     private function formatBytes(int $bytes): string
     {
         $units = ['B', 'KB', 'MB', 'GB'];
-        $pow = floor(log($bytes, 1024));
+        $pow = (int) floor(log($bytes, 1024));
         return round($bytes / (1024 ** $pow), 2) . ' ' . $units[$pow];
     }
