+++ src/Tasks/CacheMaintenanceTask.php
@@ -258,7 +258,7 @@ class CacheMaintenanceTask
         }
 
         $units = ['B', 'KB', 'MB', 'GB'];
-        $i = floor(log($bytes, 1024));
+        $i = (int) floor(log($bytes, 1024));
 
         return round($bytes / (1024 ** $i), 2) . ' ' . $units[$i];
     }
